refactor: 기물 점수 계산 책임을 Board 에서 GameResult 로 위임
